Added Options, Menu Lobby, Engine V 0.195

+++ src/SpaceBrickArena/include/Navigation.h
@@ -70,7 +70,7 @@ namespace sba
 
         int GetLineCount() const
         {
-            return ceil((this->m_LastElement + 1) / this->m_LastElement);
+            return (int)ceil((this->m_LastElement + 1) / this->m_LastElement);
         }
 
     public:
